Refactor: flyte transactional funksjoner som har med ia_sak og ia_sak_hendelse å gjøre til SamarbeidsperiodeTransactional
