+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/Transaction.kt
@@ -57,217 +57,6 @@ class Transaction(
 }
 
 context(tx: TransactionalSession)
-fun lagreHendelse(
-    hendelse: IASakshendelse,
-    sistEndretAvHendelseId: String?,
-    resulterendeStatus: IASak.Status,
-): IASakshendelse =
-    run {
-        tx.validerAtSakHarRiktigEndretAvHendelse(hendelse.saksnummer, sistEndretAvHendelseId)
-        tx.run(
-            queryOf(
-                """
-                            INSERT INTO ia_sak_hendelse (
-                                id,
-                                saksnummer,
-                                orgnr,
-                                type,
-                                resulterende_status,
-                                opprettet_av,
-                                opprettet_av_rolle,
-                                opprettet,
-                                nav_enhet_nummer,
-                                nav_enhet_navn
-                            )
-                            VALUES (
-                                :id,
-                                :saksnummer,
-                                :orgnr,
-                                :type,
-                                :resulterendeStatus,
-                                :opprettet_av,
-                                :opprettet_av_rolle,
-                                :opprettet,
-                                :enhetsnummer,
-                                :enhetsnavn
-                            ) 
-                """.trimMargin(),
-                mapOf(
-                    "id" to hendelse.id,
-                    "saksnummer" to hendelse.saksnummer,
-                    "orgnr" to hendelse.orgnummer,
-                    "type" to hendelse.hendelsesType.name,
-                    "resulterendeStatus" to resulterendeStatus.name,
-                    "opprettet_av" to hendelse.opprettetAv,
-                    "opprettet_av_rolle" to hendelse.opprettetAvRolle?.toString(),
-                    "opprettet" to hendelse.opprettetTidspunkt,
-                    "enhetsnummer" to hendelse.navEnhet.enhetsnummer,
-                    "enhetsnavn" to hendelse.navEnhet.enhetsnavn,
-                ),
-            ).asUpdate,
-        )
-        hendelse
-    }
-
-context(tx: TransactionalSession)
-fun lagreÅrsakForHendelse(
-    hendelseId: String,
-    valgtÅrsak: ValgtÅrsak,
-) = run {
-    valgtÅrsak.begrunnelser.forEach { begrunnelse ->
-        tx.run(
-            queryOf(
-                """
-                            INSERT INTO hendelse_begrunnelse (
-                                hendelse_id,
-                                aarsak,
-                                begrunnelse,
-                                aarsak_enum,
-                                begrunnelse_enum
-                            )
-                            VALUES (
-                                :hendelse_id,
-                                :aarsak,
-                                :begrunnelse,
-                                :aarsak_enum,
-                                :begrunnelse_enum
-                            ) 
-                            ON CONFLICT DO NOTHING  
-                """.trimMargin(),
-                mapOf(
-                    "hendelse_id" to hendelseId,
-                    "aarsak" to valgtÅrsak.type.navn,
-                    "begrunnelse" to begrunnelse.navn,
-                    "aarsak_enum" to valgtÅrsak.type.name,
-                    "begrunnelse_enum" to begrunnelse.name,
-                ),
-            ).asUpdate,
-        )
-    }
-}
-
-context(tx: TransactionalSession)
-fun opprettSak(iaSakDto: IASakDto): IASakDto =
-    tx.run(
-        queryOf(
-            """
-                    INSERT INTO ia_sak (
-                        saksnummer,
-                        orgnr,
-                        status,
-                        opprettet_av,
-                        opprettet,
-                        endret_av_hendelse
-                    )
-                    VALUES (
-                        :saksnummer,
-                        :orgnr,
-                        :status,
-                        :opprettet_av,
-                        :opprettet,
-                        :endret_av_hendelse
-                    )
-                    returning *                            
-            """.trimMargin(),
-            mapOf(
-                "saksnummer" to iaSakDto.saksnummer,
-                "orgnr" to iaSakDto.orgnr,
-                "status" to iaSakDto.status.name,
-                "opprettet_av" to iaSakDto.opprettetAv,
-                "opprettet" to iaSakDto.opprettetTidspunkt.toJavaLocalDateTime(),
-                "endret_av_hendelse" to iaSakDto.endretAvHendelseId,
-            ),
-        ).map { mapRowToIASakDto(it) }.asSingle,
-    )!!
-
-context(tx: TransactionalSession)
-fun oppdaterStatusPåSak(
-    saksnummer: String,
-    status: IASak.Status,
-    endretAv: String,
-    endretAvHendelseId: String,
-    oppdaterSistEndretPåSak: Boolean = true,
-): IASakDto {
-    val sistEndret: LocalDateTime = LocalDateTime.now()
-    return tx.run(
-        queryOf(
-            """
-                        UPDATE ia_sak 
-                        SET
-                            status = :status,
-                            endret_av = :endret_av,
-                            endret_av_hendelse = :endret_av_hendelse ${if (oppdaterSistEndretPåSak) ", endret = :endret" else ""}                           
-                        WHERE saksnummer = :saksnummer
-                        RETURNING *
-            """.trimMargin(),
-            mapOf(
-                "saksnummer" to saksnummer,
-                "status" to status.name,
-                "endret_av" to endretAv,
-                "endret_av_hendelse" to endretAvHendelseId,
-                "endret" to sistEndret,
-            ),
-        ).map { mapRowToIASakDto(it) }.asSingle,
-    )!!
-}
-
-private fun mapRowToIASakDto(row: Row): IASakDto = row.tilIASakDto()
-
-context(tx: TransactionalSession)
-fun hentSisteIASakDto(orgnummer: String): IASakDto? =
-    tx.run(
-        queryOf(
-            """
-            SELECT * FROM ia_sak
-            WHERE orgnr = :orgnr
-            ORDER BY opprettet DESC
-            LIMIT 1
-            """.trimIndent(),
-            mapOf("orgnr" to orgnummer),
-        ).map { mapRowToIASakDto(it) }.asSingle,
-    )
-
-context(tx: TransactionalSession)
-fun hentAlleSakerDtoForVirksomhet(orgnummer: String): List<IASakDto> =
-    tx.run(
-        queryOf(
-            """
-            SELECT * FROM ia_sak
-            WHERE orgnr = :orgnr
-            ORDER BY opprettet DESC
-            """.trimIndent(),
-            mapOf("orgnr" to orgnummer),
-        ).map { mapRowToIASakDto(it) }.asList,
-    )
-
-context(tx: TransactionalSession)
-fun settSakTilSlettet(
-    saksnummer: String,
-    hendelse: IASakshendelse,
-) {
-    tx.run(
-        queryOf(
-            """
-                        UPDATE ia_sak 
-                        SET
-                            status = :statusSlettet,
-                            endret_av = :endret_av,
-                            endret_av_hendelse = :endret_av_hendelse,
-                            endret = :endret                           
-                        WHERE saksnummer = :saksnummer
-            """.trimMargin(),
-            mapOf(
-                "saksnummer" to saksnummer,
-                "statusSlettet" to IASak.Status.SLETTET.name,
-                "endret_av" to hendelse.opprettetAv,
-                "endret_av_hendelse" to hendelse.id,
-                "endret" to hendelse.opprettetTidspunkt,
-            ),
-        ).asUpdate,
-    )
-}
-
-context(tx: TransactionalSession)
 fun hentSamarbeid(samarbeidId: Int) =
     tx.run(
         queryOf(
