+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/sideeffect/SlettSamarbeidSideEffect.kt
@@ -17,8 +17,8 @@ import no.nav.lydia.tilstandsmaskin.NyFlytService
 import no.nav.lydia.tilstandsmaskin.Transaction
 import no.nav.lydia.tilstandsmaskin.VirksomhetIATilstand
 import no.nav.lydia.tilstandsmaskin.hentSamarbeidSomIkkeErSlettet
-import no.nav.lydia.tilstandsmaskin.lagreHendelse
-import no.nav.lydia.tilstandsmaskin.oppdaterStatusPåSak
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.lagreHendelse
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.oppdaterStatusPåSak
 import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.lagreEllerOppdaterVirksomhetTilstand
 import no.nav.lydia.tilstandsmaskin.slettSamarbeid
 import java.time.LocalDateTime
