+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/sideeffect/VurderVirksomhetSideEffect.kt
@@ -17,11 +17,11 @@ import no.nav.lydia.tilgangskontroll.fia.NavAnsatt
 import no.nav.lydia.tilstandsmaskin.NyFlytService
 import no.nav.lydia.tilstandsmaskin.Transaction
 import no.nav.lydia.tilstandsmaskin.VirksomhetIATilstand
-import no.nav.lydia.tilstandsmaskin.lagreHendelse
-import no.nav.lydia.tilstandsmaskin.lagreÅrsakForHendelse
 import no.nav.lydia.tilstandsmaskin.nyHendelseBasertPåSak
-import no.nav.lydia.tilstandsmaskin.oppdaterStatusPåSak
-import no.nav.lydia.tilstandsmaskin.opprettSak
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.lagreHendelse
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.lagreÅrsakForHendelse
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.oppdaterStatusPåSak
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.SamarbeidsperiodeTransactional.Companion.opprettSak
 import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.lagreEllerOppdaterVirksomhetTilstand
 import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.slettVirksomhetTilstandAutomatiskOppdatering
